requirements.txt
easydict
python-dotenv
pytorch-toolbelt
scikit-learn
scikit-image
albumentations 
ipywidgets
python-dotenv
opencv-python
torchvision
shapely

wandb
rasterio
pycocotools
imagecodecs
imagecodecs-lite
optuna

cellpose
numpy
tqdm
pandas
